refrator util
